+++ www/whatsnew.md
@@ -12,6 +12,9 @@ Bug fixes.
 
  - Choose and reshuffle effects shuffle in the original card, rather than a copy, as intended.
  - Mind Controlled Mech now correctly cannot attack.
+ - Spiritual Diffusion now correctly destroys three random enemy minions and replaces them with 2/2 Spirits.
+ - Thelia Silentdreamer now correctly causes spells to be cast twice on her.
+ - Doby Mick's effect is now an Opener.
 
 ### 0.8.22-2.0.29 (Thursday, April 11th, 2019)